feat: 로그인 사용자의 경우 연령별 요금 할인 인수테스트 수정

+++ src/main/java/nextstep/subway/path/domain/ShortestPath.java
@@ -6,7 +6,7 @@ import nextstep.subway.station.domain.Station;
 
 public class ShortestPath {
 	private List<Station> stations;
-	private long distance;
+	private int distance;
 	private int fare;
 	private int maxLineOverFare;
 
@@ -20,7 +20,7 @@ public class ShortestPath {
 	protected ShortestPath() {
 	}
 
-	public ShortestPath(List<Station> stations, long distance, int maxLineOverFare) {
+	public ShortestPath(List<Station> stations, int distance, int maxLineOverFare) {
 		this.stations = stations;
 		this.distance = distance;
 		this.maxLineOverFare = maxLineOverFare;
